DL-17282 | Change Identifier type from Int to Long

+++ app/connectors/HipConnector.scala
@@ -47,7 +47,7 @@ class HipConnector @Inject() (
 
   private def amendProtectionUrl(
       nationalInsuranceNumber: String,
-      lifetimeAllowanceIdentifier: Int,
+      lifetimeAllowanceIdentifier: Long,
       lifetimeAllowanceSequenceNumber: Int
   ): String =
     baseUrl + s"/lifetime-allowance/person/$nationalInsuranceNumber/reference/$lifetimeAllowanceIdentifier/sequence-number/$lifetimeAllowanceSequenceNumber"
@@ -74,7 +74,7 @@ class HipConnector @Inject() (
 
   def amendProtection(
       nationalInsuranceNumber: String,
-      lifetimeAllowanceIdentifier: Int,
+      lifetimeAllowanceIdentifier: Long,
       lifetimeAllowanceSequenceNumber: Int,
       request: HipAmendProtectionRequest
   )(implicit hc: HeaderCarrier): Future[Either[UpstreamErrorResponse, HipAmendProtectionResponse]] = {
@@ -118,7 +118,7 @@ class HipConnector @Inject() (
 
   private def sendAuditEvent(
       nino: String,
-      id: Int,
+      id: Long,
       requestUrl: String,
       requestBody: HipAmendProtectionRequest,
       responseStatusCode: Int,
